initial commit. crashes on audio load
